Version 532
http://opensource.apple.com/tarballs/objc4/objc4-532.tar.gz

+++ test/ARRBase.h
@@ -9,7 +9,7 @@
 #import <Foundation/NSObject.h>
 
 @interface ARRBase : NSObject
-@property double number;
+@property long number;
 @property(retain) id object;
 @property void *pointer;
 @property(weak) __weak id delegate;
